Lunch cycle data object added
