+++ services/rule_engine/src/main/java/agents/Ava.java
@@ -4,7 +4,6 @@ import com.mindsmiths.armory.ArmoryAPI;
 import com.mindsmiths.armory.templates.BaseTemplate;
 import com.mindsmiths.emailAdapter.EmailAdapterAPI;
 import com.mindsmiths.emailAdapter.SendEmailPayload;
-import com.mindsmiths.pairingalgorithm.Days;
 import com.mindsmiths.ruleEngine.model.Agent;
 import com.mindsmiths.ruleEngine.util.DateUtil;
 import com.mindsmiths.ruleEngine.util.Log;
@@ -25,10 +24,7 @@ import java.util.*;
 @ToString(callSuper = true)
 @NoArgsConstructor
 public class Ava extends Agent {
-    private List<Days> availableDays = new ArrayList<>();
-    private String match;
     private List<String> matchHistory = new ArrayList<>();
-    private Days matchDay;
     private Date availableDaysEmailLastSentAt;
     private Date matchedWithEmailSentAt;
     private AvaLunchCycleStage lunchCycleStage;
@@ -107,12 +103,6 @@ public class Ava extends Agent {
             if (entry.getValue().getId().equals(employeeId))
                 return entry.getKey();
         return "";
-    }
-
-    public void updateAvailableDays(List<String> availableDaysStr) {
-        this.availableDays = new ArrayList<>();
-        for (String day : availableDaysStr)
-            this.availableDays.add(Days.valueOf(day));
     }
 
     public void printMatchInfo(EmployeeProfile employee, SendMatchesSignal signal) {
