+++ services/rule_engine/src/main/java/models/CmLunchCycleStage.java
@@ -2,7 +2,6 @@ package models;
 
 public enum CmLunchCycleStage {
     COLLECT_AVA_AVAILABILITIES,
-    GENERATE_MATCHES,
     COLLECT_GENERATED_MATCHES,
     SEND_TO_AVAS,
     FINISHED;
