+++ services/rule_engine/src/main/resources/rules/cultureMaster/Pairing.drl
@@ -1,6 +1,5 @@
 package rules.cultureMaster;
 
-import com.mindsmiths.ruleEngine.util.Log;
 import com.mindsmiths.ruleEngine.model.Heartbeat;
 import com.mindsmiths.ruleEngine.util.Agents;
 import com.mindsmiths.ruleEngine.model.Initialize;
@@ -13,24 +12,12 @@ import signals.EmployeeAvailabilitySignal;
 import agents.CultureMaster;
 import models.CmLunchCycleStage;
 
-
-rule "First contact"
-    salience 100
-    when
-        initialize: Initialize() from entry-point "agent-created"
-        agent: CultureMaster()
-    then
-        Log.info("Created new CultureMaster agent.");
-        delete(initialize);
-end
-
 rule "Store Employee availability"
     when
         signal: EmployeeAvailabilitySignal(employeeAvailability: employeeAvailability) from entry-point "signals"
         agent: CultureMaster(lunchCycleStage == CmLunchCycleStage.COLLECT_AVA_AVAILABILITIES)
     then
         modify(agent) {addEmployeeAvailability(employeeAvailability)};
-        Log.info("Storing availability: " + agent.getEmployeeAvailabilities());
         delete(signal);
 end
 
@@ -47,22 +34,11 @@ rule "Store Employee matching parameters"
         delete(signal);
 end
 
-rule "Generate this weeks matches"
-    when
-        Heartbeat(ts: timestamp) from entry-point "signals"
-        agent: CultureMaster(lunchCycleStage == CmLunchCycleStage.GENERATE_MATCHES)
-    then
-        Log.info("Generating matches....");
-        modify(agent) {setLunchCycleStage(CmLunchCycleStage.COLLECT_GENERATED_MATCHES)};
-        agent.generateMatches();
-end
-
 rule "Store generated matches"
     when
         generatedMatches: Matches(allMatches: allMatches) from entry-point "signals"
         agent: CultureMaster(lunchCycleStage == CmLunchCycleStage.COLLECT_GENERATED_MATCHES)
     then
-        Log.info("Matches from pairing algorithm: " + allMatches);
         modify(agent) {
             addMatches(allMatches),
             setLunchCycleStage(CmLunchCycleStage.SEND_TO_AVAS)};
@@ -82,7 +58,7 @@ end
 rule "Reset lunch cycle"
     when
         Heartbeat(ts: timestamp) from entry-point "signals"
-        agent: CultureMaster(lunchCycleStage == CmLunchCycleStage.FINISHED)
+        agent: CultureMaster(lunchCycleStage != CmLunchCycleStage.COLLECT_AVA_AVAILABILITIES)
         eval(evaluateCronExpression("* * 18-19 ? * FRI *", ts, "Europe/Zagreb"))
     then
         modify(agent) {setLunchCycleStage(CmLunchCycleStage.COLLECT_AVA_AVAILABILITIES)};
